kg-poc-app/requirements.txt
streamlit>=1.32.0
pyvis>=0.3.2
requests>=2.31.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0
sentence-transformers>=2.7.0
databricks-vectorsearch>=0.38.0
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
snowflake-connector-python>=3.8.0
anthropic>=0.27.0
networkx>=3.3
